minor: uint cast

+++ src/stress_rand_sampling.cpp
@@ -128,7 +128,7 @@ int main(int argc, const char** argv) {
     RPGraph::StressReport sr_layout1{0, layout1.graph.num_nodes(), 0, layout1.graph.num_edges(), 0};
     RPGraph::StressReport sr_layout2{0, layout2.graph.num_nodes(), 0, layout2.graph.num_edges(), 0}; // Does this zero initialize?
     std::cout << "Starting stress calculation." << std::endl;
-    while(sampled_nodes.size() < num_samples) {
+    while(sampled_nodes.size() < static_cast<uint>(num_samples)) {
         //////////////////////
         // TEMP: TIMING
         auto start = std::chrono::steady_clock::now();
